requirements.txt
transformers
sentence-transformers
faiss-cpu
pypdf
google-generativeai
docx2txt